+++ backend/stacktraces/stacktraces.go
@@ -12,7 +12,15 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-func StructureStackTrace(stackTrace string) ([]*publicModel.ErrorTrace, error) {
+type Language string
+
+const Javascript Language = "js"
+const Python Language = "python"
+const Golang Language = "golang"
+
+// StructureOTELStackTrace processes a backend opentelemetry stacktrace into a structured ErrorTraces.
+// The operation returns the deepest frame first (reversing the order of the incoming stacktrace).
+func StructureOTELStackTrace(stackTrace string) ([]*publicModel.ErrorTrace, error) {
 	jsPattern := regexp.MustCompile(` {4}at ((.+) )?\(?(.+):(\d+):(\d+)\)?`)
 	jsAnonPattern := regexp.MustCompile(` {4}at (.+) \((.+)\)`)
 	pyPattern := regexp.MustCompile(` {2}File "(.+)", line (\d+), in (\w+)`)
@@ -21,25 +29,31 @@ func StructureStackTrace(stackTrace string) ([]*publicModel.ErrorTrace, error) {
 	pyMultiPattern := regexp.MustCompile(`^During handling of the above exception, another exception occurred:$`)
 	goLinePattern := regexp.MustCompile(`\t(.+):(\d+)( 0x[0-f]+)?`)
 	goFuncPattern := regexp.MustCompile(`^(.+)\.(.+?)(\([^()]*\))?$`)
+	goRecoveredPanicPattern := regexp.MustCompile(`^\s*runtime\.gopanic\s*$`)
 	generalPattern := regexp.MustCompile(`^(.+)`)
 
-	var language string
-	var errMsg string
-	var frames []*publicModel.ErrorTrace
-	var frame *publicModel.ErrorTrace
 	var jsonStr string
 	if err := json.Unmarshal([]byte(stackTrace), &jsonStr); err == nil {
 		stackTrace = jsonStr
 	}
+	var language Language
+	var errMsg string
+	var frame *publicModel.ErrorTrace
+	frames := []*publicModel.ErrorTrace{}
 	lines := strings.Split(stackTrace, "\n")
 	for idx, line := range lines {
+		// frames explicitly set to nil means that this is part of a frame that is resetting the stacktrace
+		if frames == nil {
+			frames = []*publicModel.ErrorTrace{}
+			continue
+		}
 		if line == "Traceback (most recent call last):" {
-			language = "python"
+			language = Python
 			continue
 		}
 		if idx == 0 {
 			if line == "" {
-				language = "golang"
+				language = Golang
 				continue
 			}
 			errMsg = line
@@ -48,14 +62,14 @@ func StructureStackTrace(stackTrace string) ([]*publicModel.ErrorTrace, error) {
 		if line == "" {
 			continue
 		}
-		if language == "python" && idx == len(lines)-2 {
+		if language == Python && idx == len(lines)-2 {
 			errMsg = line
 			continue
 		}
-		if matches := pyUnderPattern.FindSubmatch([]byte(line)); language == "python" && matches != nil {
+		if matches := pyUnderPattern.FindSubmatch([]byte(line)); language == Python && matches != nil {
 			continue
 		}
-		if matches := pyMultiPattern.FindSubmatch([]byte(line)); language == "python" && matches != nil {
+		if matches := pyMultiPattern.FindSubmatch([]byte(line)); language == Python && matches != nil {
 			continue
 		}
 		if errMsg == "" {
@@ -67,7 +81,7 @@ func StructureStackTrace(stackTrace string) ([]*publicModel.ErrorTrace, error) {
 			}
 		}
 		if matches := jsPattern.FindSubmatch([]byte(line)); matches != nil {
-			language = "js"
+			language = Javascript
 			if matches[2] != nil {
 				frame.FunctionName = pointy.String(string(matches[2]))
 			}
@@ -77,30 +91,36 @@ func StructureStackTrace(stackTrace string) ([]*publicModel.ErrorTrace, error) {
 			frame.LineNumber = pointy.Int(int(l))
 			frame.ColumnNumber = pointy.Int(int(col))
 		} else if matches := jsAnonPattern.FindSubmatch([]byte(line)); matches != nil {
-			language = "js"
+			language = Javascript
 			frame.FunctionName = pointy.String(string(matches[1]))
 			frame.FileName = pointy.String(string(matches[2]))
 			frame.LineContent = pointy.String(string(matches[2]))
 		} else if matches := pyPattern.FindSubmatch([]byte(line)); matches != nil {
-			language = "python"
+			language = Python
 			frame.FunctionName = pointy.String(string(matches[3]))
 			frame.FileName = pointy.String(string(matches[1]))
 			line, _ := strconv.ParseInt(string(matches[2]), 10, 32)
 			frame.LineNumber = pointy.Int(int(line))
 			continue
+		} else if matches := goRecoveredPanicPattern.FindSubmatch([]byte(line)); matches != nil {
+			language = Golang
+			frames = nil
+			frame = nil
+			errMsg = ""
+			continue
 		} else if matches := goLinePattern.FindSubmatch([]byte(line)); matches != nil {
-			language = "golang"
+			language = Golang
 			frame.FileName = pointy.String(string(matches[1]))
 			line, _ := strconv.ParseInt(string(matches[2]), 10, 32)
 			frame.LineNumber = pointy.Int(int(line))
-		} else if matches := goFuncPattern.FindSubmatch([]byte(line)); language == "golang" && matches != nil {
+		} else if matches := goFuncPattern.FindSubmatch([]byte(line)); language == Golang && matches != nil {
 			frame.FunctionName = pointy.String(string(matches[2]))
 			continue
 		} else if matches := generalPattern.FindSubmatch([]byte(line)); matches != nil {
-			if language == "golang" {
+			if language == Golang {
 				frame.FunctionName = pointy.String(string(matches[1]))
 				continue
-			} else if language == "python" {
+			} else if language == Python {
 				if m := pyExcPattern.FindSubmatch([]byte(line)); m != nil {
 					errMsg = line
 					continue
@@ -111,11 +131,18 @@ func StructureStackTrace(stackTrace string) ([]*publicModel.ErrorTrace, error) {
 		frames = append(frames, frame)
 		frame = nil
 	}
+	// for otel non-go errors, stacktraces are sent top-down (top frame is most outer; bottom frame is most inner)
+	// our backend expects to store stack traces in the opposite order, so we have to reverse it before returning.
+	if language != Golang {
+		for i, j := 0, len(frames)-1; i < j; i, j = i+1, j-1 {
+			frames[i], frames[j] = frames[j], frames[i]
+		}
+	}
 	return frames, nil
 }
 
 func FormatStructureStackTrace(ctx context.Context, stackTrace string) string {
-	frames, err := StructureStackTrace(stackTrace)
+	frames, err := StructureOTELStackTrace(stackTrace)
 	if err != nil {
 		log.WithContext(ctx).WithField("StackTrace", stackTrace).WithError(err).Warnf("otel failed to structure stacktrace")
 		return stackTrace