+++ frontend/src/pages/Player/RightPlayerPanel/components/SessionInsights/SessionInsights.css.ts
@@ -14,3 +14,13 @@ export const insight = style({
 	},
 	width: '100%',
 })
+
+export const insightPanel = style({
+	maxHeight:
+		'calc(100vh - 160px - 32px - 42px - 32px - var(--header-height))',
+})
+
+export const insightPanelWithBanner = style({
+	maxHeight:
+		'calc(100vh - 160px - 32px - 42px - 32px - var(--header-height) - var(--banner-height))',
+})